build: generate unit tests for all groth16 mpc
